Rework the logic of TargetImages and going through image list.
All images that need to be found are now put in a linked list, with nodes pointing to the next image that needs to be found.  Cycle method, when the location of the given image is found, goes to the next node and repeats the process. Each TargetImage object can have multiple nodes, which takes into account different lokoing images that need to be clicked for the same step. The Step function will go through each node until it finds a match.