Rework writing images to database
Images can now be written in the database by a single function WriteTargetImagesToDatabase, which takes in a path to a directory where the images are found. Images must be named after a specific convention (image number followed by a list of next nodes in brackets, divided by  ',' if there are more nodes, after the brackets [] true if image is before an ad.